// mesh-control-plane/mesh-ui/src/api/openapi/types.gen.ts
export type V1User = {
  id?: string
  name?: string
  createdAt?: string
  displayName?: string
  email?: string
  providerId?: string
  provider?: string
  profilePicUrl?: string
}